Extract URLs into separate helper
